Simplified the folder structure for the physics system
