+++ atlas/drivers/jolt-cpp/jolt_context.hpp
@@ -3,7 +3,7 @@
 #include <drivers/jolt-cpp/jolt_contact_listener.hpp>
 #include <core/core.hpp>
 
-#include <physics/physics_3d/physics_context.hpp>
+#include <physics/physics_context.hpp>
 #include <drivers/jolt-cpp/jolt_components.hpp>
 #include <drivers/jolt-cpp/interface/jolt_broad_phase.hpp>
 #include <physics/components.hpp>
