+++ src/atlas/physics/physics_context.cpp
@@ -1,13 +1,12 @@
+#include <type_traits>
+#include <Jolt/Jolt.h>
 #include <physics/physics_context.hpp>
-#include <drivers/jolt-cpp-experiemental/jolt_context.hpp>
+#include <drivers/jolt-cpp/jolt_context.hpp>
 
-namespace atlas {
-    ref<physics_context> initialize_physics_context(const physics_settings& p_settings) {
-        switch (p_settings.api_context) {
-        case physics_api::jolt:
-            return create_ref<physics::experiemental::jolt_context>(p_settings);
-        }
+namespace atlas::physics {
 
-        assert(false);
+    ref<physics_context> initialize_physics_context(const jolt_settings& p_settings) {
+        return create_ref<jolt_context>(p_settings);
     }
-};
+
+}
