+++ tests/jolt_engine.test.cpp
@@ -13,7 +13,7 @@
 
 #include <drivers/jolt-cpp/jolt_helper.hpp>
 #include <drivers/jolt-cpp/jolt_components.hpp>
-#include <physics/physics_3d/physics.hpp>
+#include <physics/physics.hpp>
 
 namespace atlas {
